+++ NetIdMobileSdk/NetIdMobileSdk-App/ServiceViewModel.swift
@@ -22,6 +22,8 @@ class ServiceViewModel: NSObject, ObservableObject {
     @Published var authenticationEnabled = false
     @Published var userInfoEnabled = false
     @Published var endSessionEnabled = false
+    @Published var fetchPermissionsEnabled = false
+    @Published var updatePermissionEnabled = false
 
     @Published var authorizationViewVisible = false
 
@@ -29,7 +31,7 @@ class ServiceViewModel: NSObject, ObservableObject {
     @Published var authenticationStatusColor = Color.gray
     @Published var userInfoStatusColor = Color.gray
 
-    @Published var logText = "Logs:\n\n"
+    @Published var logText = ""
 
     func initializeNetIdService() {
         initializationEnabled = false
@@ -49,6 +51,19 @@ class ServiceViewModel: NSObject, ObservableObject {
     func fetchUserInfo() {
         userInfoEnabled = false
         NetIdService.sharedInstance.fetchUserInfo()
+    }
+
+    func fetchPermissions() {
+        fetchPermissionsEnabled = false
+        NetIdService.sharedInstance.fetchPermissions(collapseSyncId: false)
+    }
+
+    func updatePermission() {
+        updatePermissionEnabled = false
+        // these values are only for demonstration purpose
+        NetIdService.sharedInstance.updatePermission(NetIdPermissionUpdate(idConsent: "VALID",
+                iabTc: "CPdfZIAPdfZIACnABCDECbCkAP_AAAAAAAYgIzJd9D7dbXFDefx_SPt0OYwW0NBXCuQCChSAA2AFVAOQcLQA02EaMATAhiACEQIAolIBAAEEHAFEAECQQIAEAAHsAgSEhAAKIAJEEBEQAAIQAAoKAAAAAAAIgAABoASAmBiQS5bmRUCAOIAQRgBIgggBCIADAgMBBEAIABgIAIIIgSgAAQAAAKIAAAAAARAAAASGgFABcAEMAPwAgoBaQEiAJ2AUiAxgBnwqASAEMAJgAXABHAEcALSAkEBeYDPh0EIABYAFQAMgAcgA-AEAALgAZAA0AB4AD6AIYAigBMACfAFwAXQAxABmADeAHMAPwAhgBLACYAE0AKMAUoAsQBbgDDAGiAPaAfgB-gEDAIoARaAjgCOgEpALEAWmAuYC6gF5AMUAbQA3ABxADnAHUAPQAi8BIICRAE7AKHAXmAwYBjADJAGVAMsAZmAz4BrADiwHjgPrAg0BDkhAbAAWABkAFwAQwAmABcADEAGYAN4AjgBSgCxAIoARwAlIBaQC5gGKANoAc4A6gB6AEggJEAScAz4B45KBAAAgABYAGQAOAAfAB4AEQAJgAXAAxABmADaAIYARwAowBSgC3AH4ARwAk4BaQC6gGKANwAdQBF4CRAF5gMsAZ8A1gCGoSBeAAgABYAFQAMgAcgA8AEAAMgAaAA8gCGAIoATAAngBvADmAH4AQgAhgBHACWAE0AKUAW4AwwB7QD8AP0AgYBFICNAI4ASkAuYBigDaAG4AOIAegBIgCdgFDgKRAXmAwYBkgDPoGsAayA4IB44EOREAYAQwA_AEiAJ2AUiAz4ZAHACGAEwARwBHAEnALzAZ8UgXAALAAqABkADkAHwAgABkADQAHkAQwBFACYAE8AKQAYgAzABzAD8AIYAUYApQBYgC3AGjAPwA_QCLQEcAR0AlIBcwC8gGKANoAbgA9ACLwEiAJOATsAocBeYDGAGSAMsAZ9A1gDWQHBAPHAhm.f_gAAAAAAsgA"),
+                collapseSyncId: false)
     }
 
     func endSession() {
@@ -86,6 +101,8 @@ extension ServiceViewModel: NetIdServiceDelegate {
         authenticationStatusColor = Color.green
         userInfoEnabled = true
         endSessionEnabled = true
+        updatePermissionEnabled = true
+        fetchPermissionsEnabled = true
         logText.append("Net ID service authorized successfully\n" + accessToken + "\n")
     }
 
@@ -122,6 +139,8 @@ extension ServiceViewModel: NetIdServiceDelegate {
         userInfoStatusColor = Color.gray
         authenticationEnabled = true
         userInfoEnabled = false
+        updatePermissionEnabled = false
+        fetchPermissionsEnabled = false
     }
 
     func didEncounterNetworkError(_ error: NetIdError) {
@@ -131,6 +150,7 @@ extension ServiceViewModel: NetIdServiceDelegate {
                 preferredStyle: .alert)
         alert.addAction(UIAlertAction(title: NSLocalizedString("network_error_alert_action", comment: ""), style: .default) { _ in
             alert.dismiss(animated: true)
+
             switch error.process {
             case .Configuration:
                 self.initializationStatusColor = Color.red
@@ -141,6 +161,12 @@ extension ServiceViewModel: NetIdServiceDelegate {
             case .UserInfo:
                 self.userInfoStatusColor = Color.red
                 self.userInfoEnabled = true
+            case .PermissionRead:
+                //TODO
+                self.logText.append("")
+            case .PermissionWrite:
+                //TODO
+                self.logText.append("")
             }
         })
         UIApplication.shared.visibleViewController?.present(alert, animated: true)
@@ -161,6 +187,32 @@ extension ServiceViewModel: NetIdServiceDelegate {
         case .UserInfo:
             userInfoStatusColor = Color.yellow
             userInfoEnabled = true
+        case .PermissionRead:
+            //TODO
+            logText.append("")
+        case .PermissionWrite:
+            //TODO
+            logText.append("")
         }
+    }
+
+    public func didFetchPermissions(_ permissions: Permissions) {
+        logText.append("didFetchPermissions \(permissions.description) \n")
+        fetchPermissionsEnabled = true
+    }
+
+    public func didFetchPermissionsWithError(_ error: NetIdError) {
+        logText.append("didFetchPermissionsWithError \(error.code.rawValue)\n")
+        fetchPermissionsEnabled = true
+    }
+
+    public func didUpdatePermission() {
+        logText.append("didUpdatePermission \n")
+        updatePermissionEnabled = true
+    }
+
+    public func didUpdatePermissionWithError(_ error: NetIdError) {
+        logText.append("didUpdatePermissionWithError \(error.code.rawValue)\n")
+        updatePermissionEnabled = true
     }
 }