+++ NetIdMobileSdk/NetIdMobileSdk/api/NetIdConfig.swift
@@ -15,15 +15,15 @@
 import Foundation
 
 public struct NetIdConfig {
+    public var host: String
+    public var clientId: String
+    public var redirectUri: String
+    public var originUrlScheme: String
+
     public init(host: String, clientId: String, redirectUri: String, originUrlScheme: String) {
         self.host = host
         self.clientId = clientId
         self.redirectUri = redirectUri
         self.originUrlScheme = originUrlScheme
     }
-
-    public var host: String
-    public var clientId: String
-    public var redirectUri: String
-    public var originUrlScheme: String
 }