+++ NetIdMobileSdk/NetIdMobileSdk/manager/UserInfoManager.swift
@@ -1,0 +1,44 @@
+// Copyright 2022 European netID Foundation (https://enid.foundation)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import Foundation
+
+class UserInfoManager: NSObject {
+    private weak var delegate: UserInfoManagerDelegate?
+
+    init(delegate: UserInfoManagerDelegate?) {
+        self.delegate = delegate
+        super.init()
+    }
+
+    /**
+     Fetches the UserInfo.
+     */
+    public func fetchUserInfo(host: String, accessToken: String) {
+        let userInfoRequest = UserInfoRequest(host: host, accessToken: accessToken)
+        Webservice.shared.performRequest(userInfoRequest, callback: { data, error in
+            guard let data = data else {
+                self.delegate?.didFetchUserInfoWithError(NetIdError(code: .Unknown, process: .UserInfo)
+                )
+                return
+            }
+
+            if let userInfo = try? JSONDecoder().decode(UserInfo.self, from: data) {
+                self.delegate?.didFetchUserInfo(userInfo)
+            } else {
+                self.delegate?.didFetchUserInfoWithError(NetIdError(code: .JsonDeserializationError, process: .UserInfo))
+            }
+        })
+    }
+}