+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AppAuthManager for the library authentication communication  
 - Authorization UI 
 - Necessary App2App utilities 
+- Permission management webservices
+- Permission management API functions
 
 ### Fixed
 - Url schemes error